member 타입 long으로 변경

+++ insider/src/main/java/com/kh/insider/dto/DmMessageDto.java
@@ -7,7 +7,7 @@ import lombok.Data;
 @Data
 public class DmMessageDto {
 
-	private int messageNo;
+	private long messageNo;
 	private int roomNo;
 	private int messageSender;
 	private String messageContent;
